requirements.txt
selenium==4.24.0
beautifulsoup4
praw~=7.7.1
python-dotenv~=1.0.1
openai~=1.44.1
pysrt~=1.1.2
